backend/pyproject.toml
[project]
name = "bad-guys-backend"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
  "Django==5.2.*",
  "djangorestframework>=3.15",
  "drf-nested-routers>=0.93",
  "psycopg[binary]>=3.2",
  "django-cors-headers>=4.4",
  "djangorestframework-simplejwt>=5.3",
  "redis>=5.0",
  "celery>=5.4",
  "httpx>=0.27",
  "weasyprint>=61.0",
  "uvicorn>=0.30",
  "gunicorn>=22.0",
  "python-dotenv>=1.0",
  "django-fsm==2.8.*",
  "drf-spectacular>=0.27",
  "django-filter>=24.2",
  "prometheus-client>=0.20",
  "django-allauth>=64.0"
  , "requests>=2.32"
  , "cryptography>=42.0"
]

[tool.uv]
dev-dependencies = ["pytest", "pytest-django", "pytest-cov", "ruff", "mypy"]
